install sh

+++ install/centos/centos_php.sh
@@ -7,7 +7,9 @@ yum install unzip -y
 #######新建php用户和php组
 groupadd -r php && useradd -r -g php -s /bin/false -d /usr/local/php7 -M php
 ######从GitHub下载php7安装包
-wget -c --no-check-certificate -O php7-src-master.zip https://github.com/php/php-src/archive/master.zip
+if [ ! -f "php7-src-master.zip" ]; then
+  wget -c --no-check-certificate -O php7-src-master.zip https://github.com/php/php-src/archive/master.zip
+fi
 ######开始解压php7包
 unzip -q php7-src-master.zip && cd php-src-master
 #####安装编译php7时需要的依赖包
